requirements.txt
streamlit
pandas
numpy
joblib
plotly
pinecone-client
sentence-transformers
transformers
pymupdf
